+++ tests/test_epoch_mapping.py
@@ -9,7 +9,7 @@ import unittest
 import numpy as np
 from tests.recording_data import minimal_recording_data
 
-from twopy import interpolate_stimulus_epochs_to_frame_windows
+from twopy.analysis.epoch_mapping import interpolate_stimulus_epochs_to_frame_windows
 from twopy.converted import RecordingData